pyproject.toml
[tool.poetry]
name = "transcribe-meetings"
version = "0.1.0"
description = "Transcribe meeting audio and generate actionable summaries"
authors = ["Dhruv Nigam"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11,<3.13"
openai-whisper = "^20231117"
torch = "^2.1.0,<2.5.0"
dspy-ai = "^2.4.0"
loguru = "^0.7.2"
python-dotenv = "^1.0.0"
click = "^8.1.7"
pydantic = "^2.5.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.12.0"
mypy = "^1.7.0"
ruff = "^0.1.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


